moved errors into a seeperate file
