+++ pysondb/db.py
@@ -7,6 +7,11 @@ import uuid
 from pathlib import Path
 from typing import Any, Callable, Dict, List, Union, Pattern
 
+#errors
+from pysondb.errors.db_errors import IdNotFoundError
+from pysondb.errors.db_errors import DataNotFoundError
+from pysondb.errors.db_errors import SchemaError
+
 from filelock import FileLock
 
 # constants
@@ -16,46 +21,6 @@ EMPTY_DATA: Dict[str, Any] = {"data": []}
 # logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger("pysondb")
 logger.setLevel(logging.DEBUG)
-
-
-# Errors
-class DataNotFoundError(Exception):
-    """Exception raised if id not found.
-
-    Attributes:
-        data
-    """
-
-    def __init__(self, data: Dict[str, Any]) -> None:
-        self.data = data
-
-    def __str__(self) -> str:
-        return f"The data {self.data!r} does not exists in JSON db"
-
-
-class IdNotFoundError(Exception):
-    """Exception raised if id not found.
-
-    Attributes:
-        pk -- primary key / id
-    """
-
-    def __init__(self, pk: int) -> None:
-        self.pk = pk
-
-    def __str__(self) -> str:
-        return f"Id {self.pk!r} does not exist in the JSON db"
-
-
-class SchemaError(Exception):
-    """Exception raised for field/key errors."""
-
-    def __init__(self, *args) -> None:
-        self.args = args
-
-    def __str__(self) -> str:
-        return str(self.args)
-
 
 # util functions
 def create_db(filename: str, create_file: bool = True) -> True:
